Controlador de User movido a src/controller/user
